🚚rename: add api folder
Related to: #73, #74
